# Conflicts:
#	target/m2e-wtp/web-resources/META-INF/maven/com.team3/mountainBooks/pom.properties

+++ src/main/java/com/team3/service/Service.java
@@ -1,6 +1,8 @@
 package com.team3.service;
 
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.*;
 import java.text.*;
 import javax.servlet.http.*;
@@ -44,7 +46,10 @@ import com.team3.user.faq.dao.FaqDao;
 import com.team3.user.faq.dto.FaqDto;
 import com.team3.user.interest.dao.InterestDao;
 import com.team3.user.interest.dto.InterestDto;
-import com.team3.user.map.dao.MapDao;
+import com.team3.user.map.dao.PaymentDao;
+import com.team3.user.map.dto.PaymentPointDto;
+import com.team3.user.map.dto.PointDto;
+import com.team3.user.member.dto.MemberAddressDto;
 import com.team3.user.member.dto.MemberDto;
 import com.team3.user.order.dao.OrderDao;
 import com.team3.user.order.dto.CartDto;
@@ -85,13 +90,13 @@ public class Service implements ServiceInterface {
 	private AdminMapDao adminMapDao;
 
 	@Autowired
+	private PaymentDao paymentDao;
+	
+	@Autowired
 	private BookDao bookDao;
 
 	@Autowired
 	private MemberManageDao memberManageDao;
-
-	@Autowired
-	private MapDao mapDao;
 
 	@Autowired
 	private FaqDao faqDao;
@@ -395,19 +400,26 @@ public class Service implements ServiceInterface {
 		LogAspect.logger.info(LogAspect.logMsg + check);
 
 		mav.addObject("check", check);
-
 		mav.setViewName("adminFaqInsertOk.admin");
 	}
 
+
+	/*
+	 * public void deleteFile(MapDto mapDto,String[] oldPathList) { String realPath
+	 * = Service.class.getResource("").getPath()
+	 * .replace("apache-tomcat-8.0.47/wtpwebapps", "workspace")
+	 * .replace("WEB-INF/classes/com/team3/service/", "src/main/webapp");
+	 * 
+	 * for(int i=0;i<oldPathList.length;i++) { File file=new File(realPath+
+	 * "/adminImg",oldPathList[i]); file.delete(); } }
+	 */
 	@Override
 	public void adminFaqMain(ModelAndView mav) {
 		Map<String, Object> map = mav.getModelMap();
 		HttpServletRequest request = (HttpServletRequest) map.get("request");
 
 		String pageNumber = request.getParameter("pageNumber");
-		if (pageNumber == null)
-			pageNumber = "1";
-
+		if (pageNumber == null)pageNumber = "1";
 		int currentPage = Integer.parseInt(pageNumber);
 
 		int count = adminFaqDao.faqCount();
@@ -753,7 +765,7 @@ public class Service implements ServiceInterface {
 
 		check = adminMapDao.mapUpdate(mapDto);
 		LogAspect.logger.info(LogAspect.logMsg + check);
-		if (check > 0) {
+		if (check > 0&&mapDto.getImg_path()!=null) {
 			deleteFile(oldPathList);
 		}
 
@@ -1425,8 +1437,66 @@ public class Service implements ServiceInterface {
 			}
 		}
 	}
-
+	
 	@Override
+	public void payment(ModelAndView mav) {
+		Map<String, Object> map = mav.getModelMap();
+		HttpServletRequest request=(HttpServletRequest) map.get("request");
+		
+		String isbn=request.getParameter("isbn");
+		
+		String[] isbnList=isbn.split(",");
+		String trantIsbn="";
+		
+		for(int i=0;i<isbnList.length;i++) {
+			trantIsbn+=isbnList[i];
+		}
+		
+		HttpSession session = request.getSession(); //세션받기 ID
+		MemberDto memberDto=memberDao.updateAccount(session);
+		
+		LogAspect.logger.info(LogAspect.logMsg +isbnList.length);
+		
+		if(request.getParameter("val").equals("1")) {
+			int quantity=Integer.parseInt(request.getParameter("quantity").split(",")[0]);
+			LogAspect.logger.info(LogAspect.logMsg +isbn+"\t"+quantity);
+			
+			BookDto bookDto=paymentDao.selectBook(isbnList[0]);
+			mav.addObject("bookDto",bookDto);
+			mav.addObject("quantity",quantity);
+			mav.addObject("length",isbnList.length);
+			mav.addObject("isbn",isbn);
+		}else if(request.getParameter("val").equals("2")) {
+			String quantity=request.getParameter("quantity");
+			String[] quantityList=quantity.split(",");
+			String transQuantity="";
+			for(int i=0;i<quantityList.length;i++) {
+				transQuantity+=quantityList[i]+"/";
+			}
+			BookDto firstBookDto=new BookDto();
+			int count=0;
+			long price=0;
+			List<BookDto>bookList=new ArrayList<BookDto>();
+			for(int i=0;i<isbnList.length;i++) {
+				count+=Integer.parseInt(quantityList[i]);
+				BookDto bookDto=paymentDao.selectBook(isbnList[i]);
+				if(i==0) {
+					firstBookDto=bookDto;
+				}
+				price+=Integer.parseInt(quantityList[i])* bookDto.getPrice();
+				bookList.add(bookDto);
+			}
+			mav.addObject("quantity",transQuantity);
+			mav.addObject("isbn",trantIsbn);
+			mav.addObject("price",price);
+			mav.addObject("bookListSize",bookList.size()-1);
+			mav.addObject("bookDto",firstBookDto);
+			mav.addObject("count",count);
+		}
+		mav.addObject("memberDto",memberDto);
+		mav.setViewName("payment.users");
+	}
+
 	public void getTopTen(ModelAndView mav) {
 		List<FaqDto> faqDtoTTList = faqDao.getTopTenList();
 		mav.addObject("faqDtoTTList", faqDtoTTList);
@@ -1436,7 +1506,7 @@ public class Service implements ServiceInterface {
 	@Override
 	public void getFaq(ModelAndView mav) {
 		List<FaqDto> faqDtoList = faqDao.getTopTenList();
-
+		
 		for (int i = 0; i < faqDtoList.size(); i++) {
 			faqDtoList.get(i).setContent(faqDtoList.get(i).getContent().replace("\r\n", "<br />"));
 		}
@@ -1446,6 +1516,102 @@ public class Service implements ServiceInterface {
 	}
 
 	@Override
+	public void paymentOk(ModelAndView mav) {
+		Map<String, Object> map = mav.getModelMap();
+		HttpServletRequest request=(HttpServletRequest) map.get("request");
+		
+		int flag=0;
+		if(request.getParameter("flag")!=null)
+			flag=Integer.parseInt(request.getParameter("flag"));
+		
+		PaymentPointDto paymentPointDto=(PaymentPointDto)map.get("paymentPointDto");
+		OrderDto orderDto=(OrderDto) map.get("orderDto");
+		
+		HttpSession session=request.getSession();
+		MemberDto memberDto=memberDao.updateAccount(session);
+		
+		String order_number=Long.toString(System.currentTimeMillis())+"_"+memberDto.getMember_number();
+		paymentPointDto.setOrder_number(order_number);
+		orderDto.setOrder_number(order_number);
+		orderDto.setId(memberDto.getId());
+		
+		if(paymentPointDto.getDirect_deposit()==null||paymentPointDto.getDirect_deposit().equals("")) {
+			orderDto.setOrder_status(1);
+		}
+		
+		LogAspect.logger.info(LogAspect.logMsg+paymentPointDto.toString());
+		LogAspect.logger.info(LogAspect.logMsg+orderDto.toString());
+		
+		HashMap<String, Object>hmap=new HashMap<String, Object>();
+		hmap.put("total_price", orderDto.getTotal_price());
+		hmap.put("point_history",paymentPointDto.getPoint_history());
+		hmap.put("sales_total", orderDto.getTotal_price()+paymentPointDto.getPoint_history());
+		hmap.put("save_point",(paymentPointDto.getSave_point()-paymentPointDto.getPoint_history()));
+		hmap.put("id",orderDto.getId());
+		hmap.put("goods",orderDto.getGoods());
+		
+		String[] goods=((String)hmap.get("goods")).split("/");
+		if(flag==1)
+			LogAspect.logger.info(LogAspect.logMsg+goods.length);
+		
+		HashMap<String, Object> cartMap=new HashMap<String, Object>();
+		ArrayList<String>isbnList=null;
+		if(flag==1) {
+			isbnList=new ArrayList<String>();
+			for(int i=0;i<goods.length;i++) {
+				isbnList.add(goods[i]+"/");
+			}
+		}
+		cartMap.put("id", orderDto.getId());
+		cartMap.put("isbnList", isbnList);
+		
+		int check=paymentDao.paymentOk(paymentPointDto,orderDto,hmap,cartMap);
+		LogAspect.logger.info(LogAspect.logMsg+"paymentOk check : "+check);
+		mav.addObject("check",check);
+		mav.setViewName("paymentOk.users");
+	}
+
+	@Override
+	public void addressList(ModelAndView mav) {
+		Map<String, Object> map = mav.getModelMap();
+		HttpServletRequest request = (HttpServletRequest) map.get("request");
+
+		HttpSession session = request.getSession();
+		String id = (String) session.getAttribute("mbId");
+
+		List<MemberAddressDto> memberAddressDtoList = paymentDao.getMemberAddress(id);
+		mav.addObject("memberAddressDtoList", memberAddressDtoList);
+		mav.addObject("size", memberAddressDtoList.size());
+		mav.setViewName("addressList.empty");
+	}
+
+	@Override
+	public void addAddress(ModelAndView mav) {
+		Map<String, Object> map = mav.getModelMap();
+		HttpServletRequest request = (HttpServletRequest) map.get("request");
+		MemberAddressDto memberAddressDto = (MemberAddressDto) map.get("memberAddressDto");
+		String id = (String) request.getSession().getAttribute("id");
+		memberAddressDto.setId(id);
+
+		int check = paymentDao.insertZipcode(memberAddressDto);
+
+		mav.addObject("check", check);
+		mav.setViewName("addAddress.empty");
+	}
+
+	@Override
+	public void deleteAddress(ModelAndView mav) {
+		Map<String, Object> map = mav.getModelMap();
+		HttpServletRequest request = (HttpServletRequest) map.get("request");
+		String member_zipcode = request.getParameter("check");
+
+		int check = paymentDao.deleteMemeberAddress(member_zipcode);
+		LogAspect.logger.info(LogAspect.logMsg + "check : " + check);
+
+		mav.addObject("check", check);
+		mav.setViewName("addressDelete.empty");
+	}
+
 	public void bookList(ModelAndView mav) {
 		Map<String, Object> map = mav.getModelMap();
 		HttpServletRequest request = (HttpServletRequest) map.get("request");
@@ -1650,9 +1816,7 @@ public class Service implements ServiceInterface {
 	public void adminBookSearch(ModelAndView mav) {
 		Map<String, Object> map = mav.getModelMap();
 		HttpServletRequest request = (HttpServletRequest) map.get("request");
-
 		List<BookDto> bookList = adminBook.getAdminBookSearch();
-
 		mav.addObject("bookList", bookList);
 	}
 
@@ -2653,5 +2817,36 @@ public class Service implements ServiceInterface {
 //			mav.addObject("list_id", list_id);
 //			mav.addObject("check", Integer.parseInt(check));
 		}
+	}
+	
+	@Override
+	public void userPoint(ModelAndView mav) {
+		Map<String, Object> map=mav.getModelMap();
+		HttpServletRequest request=(HttpServletRequest) map.get("request");
+		HttpSession session = request.getSession();
+		int point=0;
+		if(session.getAttribute("mbId")!=null) {
+			String id=(String) session.getAttribute("mbId");
+			int check=1;
+			List<PointDto>list=paymentDao.selectPoint(id);
+			SimpleDateFormat sdf=new SimpleDateFormat("yyyy-MM-dd");
+			for(int i=0;i<list.size();i++) {
+				list.get(i).setStr_order_date(sdf.format(list.get(i).getOrder_date()));
+			}
+			point=(int)list.get(0).getPoint();
+			
+			HashMap<String, Object> selectMap=paymentDao.selectState(id);
+			
+			mav.addObject("check",check);
+			mav.addObject("point",point);
+			mav.addObject("list",list);
+			mav.addObject("selectMap",selectMap);
+			mav.setViewName("userPointView.users");
+		}else {
+			int check=0;
+			mav.addObject("check",check);
+			mav.setViewName("userPointView.users");
+		}
+		
 	}
 }
